requirements.txt
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
autopep8
certifi
cffi
chardet
charset-normalizer
colorama
cryptography
defusedxml
distro
Django
django-cors-headers
django-countries
django-crispy-forms
django-debug-toolbar
django-environ
django-filter
django-templated-mail
django-tinymce
djangorestframework
djangorestframework-simplejwt
djoser
filelock
frozenlist
h11
httpcore
httpx
idna
importlib_metadata
Markdown
multidict
oauthlib
openai
pep8
pillow
platformdirs
psycopg2-binary
pycodestyle
pycparser
pydantic
pydantic_core
PyJWT
python-decouple
python3-openid
pytz
requests
requests-oauthlib
sniffio
social-auth-app-django
social-auth-core
sqlparse
tqdm
typing_extensions
tzdata
urllib3
wheel
yarl
zipp
